AD IdP: consolidate WSO2 and AD FS code
They now use the same code, but only differ in the templates for the
soap request.
